checking admin functionality

+++ lib/live_email_notification_web/live/app/user_live.ex
@@ -12,15 +12,16 @@ defmodule LiveEmailNotificationWeb.UserLive do
       <div class="mx-auto px-4 py-6 sm:px-6 space-y-1 lg:px-8">
         <div>
           <h1 class="text-2xl font-bold tracking-tight text-gray-900 capitalise relative">
-            <.link href={~p"/users/settings"} class="text-brand"><%= @current_user.first_name <> "'s" %></.link> <%= String.split(@current_path, "/") %>
+            Users
           </h1>
-          <p class="text-sm text-slate-500 hover:text-slate-600">View users and do <%= @current_user.user_type.user_type <> "'s" %> stuff.</p>
+          <p class="text-sm text-slate-500 hover:text-slate-600">View users and do <span class="border-b border-brand"><%= @current_user.user_type.user_type %></span> stuff.</p>
         </div>
         <.table id="users" rows={@users}>
           <:col :let={user} label="Full Name"><%= user.first_name %> <%= user.last_name %></:col>
           <:col :let={user} label="Email"><%= user.email %></:col>
           <:col :let={user} label="Phone no."><%= user.msisdn %></:col>
           <:col :let={user} label="Plan"><%= user.plan.plan_name %></:col>
+          <:col :let={user} label="Permission"><%= user.user_type.user_type %></:col>
           <:col :let={user} label="Actions">
             <div class="space-x-1">
               <button phx-click="view" phx-value-selected={user.uuid} phx-value-context="view" type="button" class="border bg-blue-50 p-0.5 cursor-pointer has-tooltip">
@@ -66,7 +67,7 @@ defmodule LiveEmailNotificationWeb.UserLive do
           :if={@modal_active}
           id="user-dialogue"
           show
-          on_cancel={JS.navigate(~p"/users")}
+          on_cancel={JS.navigate(~p"/admin/users")}
           class={[@modal_context == "delete" && "!p-1"]}
         >
           <.live_component
@@ -91,8 +92,6 @@ defmodule LiveEmailNotificationWeb.UserLive do
       |> assign(
           users: []
       )
-      |> put_flash(:error, "You shouldn't be here.")
-      |> redirect(to: ~p"/dashboard")
     socket
   end
 
@@ -100,7 +99,7 @@ defmodule LiveEmailNotificationWeb.UserLive do
     socket =
       socket
       |> assign(
-           users: User |> where([user], user.id) |> Repo.all() |> Repo.preload([:plan])
+           users: Repo.all(from(User)) |> Repo.preload([:user_type, :plan])
          )
     socket
   end
@@ -109,44 +108,36 @@ defmodule LiveEmailNotificationWeb.UserLive do
     socket =
       socket
       |> assign(
-           users: User |> where([user], user.id) |> Repo.all() |> Repo.preload([:plan])
+           users: Repo.all(from(User)) |> Repo.preload([:user_type, :plan])
          )
     socket
   end
 
   def mount(params, _session, socket) do
-    # if superuser show all users
-    # if admin show customers
-    # if basic user redirect to dashboard and flash un-authorised
-    if user = socket.assigns.current_user do
-      socket = case user.user_type.user_type do
-        "superuser" ->
-          mount_super(socket)
-        "admin" ->
-          mount_admin(socket)
-        "user" ->
-          mount_basic(socket)
-        _ ->
-          mount_basic(socket)
-      end
-
-      {:ok, socket
-        |> assign(
-             page_title: "Users",
-             selected_user: nil,
-             check_errors: false,
-             modal_active: Map.has_key?(params, "action"),
-             modal_context: Map.get(params, "action"),
-             custom_error: nil,
-             plans: Plan |> where([plan], plan.id) |> Repo.all()
-           )
-      }
-    else
-      socket = socket
-                |> put_flash(:error, "You shouldn't be here.")
-               |> redirect(to: ~p"/dashboard")
-      {:ok, assign(socket, page_title: "Users")}
+    socket = case socket.assigns.current_user.user_type.user_type do
+      "superuser" ->
+        mount_super(socket)
+      "admin" ->
+        mount_admin(socket)
+      "user" ->
+        mount_basic(socket)
+      _ ->
+        mount_basic(socket)
     end
+
+    {
+      :ok,
+      socket
+      |> assign(
+         page_title: "Users",
+         selected_user: nil,
+         check_errors: false,
+         modal_active: Map.has_key?(params, "action"),
+         modal_context: Map.get(params, "action"),
+         custom_error: nil,
+         plans: Repo.all(from(Plan))
+      )
+    }
   end
 
   def handle_event("showModal", %{ "selected" => user_id, "context" => context }, socket) do
@@ -161,19 +152,40 @@ defmodule LiveEmailNotificationWeb.UserLive do
 
   def handle_event("view", %{ "selected" => uuid }, socket) do
     socket = socket
-             |> redirect(to: ~p"/users/#{uuid}/dashboard")
+             |> redirect(to: "/admin/users/#{uuid}/dashboard")
     {:noreply, socket}
   end
 
   def handle_event("delete", %{ "selected" => user_id}, socket) do
     try do
       us_id = elem(Integer.parse(user_id), 0)
-      Repo.delete_all from(user in User, where: user.id == ^us_id)
-      socket = socket
-               |> assign(:modal_active, !socket.assigns.modal_active)
-               |> put_flash(:info, "User deleted successfully.")
-               |> redirect(to: ~p"/users")
-      {:noreply, socket}
+      if socket.assigns.selected_user.user_type.user_type == "superuser" || socket.assigns.selected_user.user_type.user_type == "admin" do
+        if socket.assigns.current_user.user_type.user_type == "superuser" do
+          if socket.assigns.selected_user.user_type.user_type == "superuser" do
+            socket = socket |> put_flash(:error, "You are not allowed to perform this action.")
+                     |> assign(:modal_active, !socket.assigns.modal_active)
+                     |> redirect(to: ~p"/admin/users")
+            {:noreply, socket}
+          else
+            Repo.delete_all from(user in User, where: user.id == ^us_id)
+            socket = socket |> put_flash(:info, "User deleted successfully.")
+                     |> assign(:modal_active, !socket.assigns.modal_active)
+                     |> redirect(to: ~p"/admin/users")
+            {:noreply, socket}
+          end
+        else
+          socket = socket |> put_flash(:error, "You are not allowed to perform this action.")
+                   |> assign(:modal_active, !socket.assigns.modal_active)
+                   |> redirect(to: ~p"/admin/users")
+          {:noreply, socket}
+        end
+      else
+        Repo.delete_all from(user in User, where: user.id == ^us_id)
+        socket = socket |> put_flash(:info, "User deleted successfully.")
+                 |> assign(:modal_active, !socket.assigns.modal_active)
+                 |> redirect(to: ~p"/admin/users")
+        {:noreply, socket}
+      end
     catch
       message ->
         socket = socket
